requirements.txt
python-telegram-bot==20.0
configparser==5.3.0
selenium~=4.17.2
beautifulsoup4~=4.12.3